move storage type to feed managaer
