+++ RUFAS/routines/feed_storage/enums.py
@@ -79,43 +79,3 @@ class CropType(Enum):
     FINE_FESCUE = "Fine Fescue"
     TALL_FESCUE = "Tall Fescue"
     MEADOW_FESCUE = "Meadow Fescue"
-
-
-class StorageType(Enum):
-    """
-    Enumeration of storage types.
-
-    Attributes
-    ----------
-    PROTECTED_INDOORS : str
-        Represents a type of hay storage.
-    PROTECTED_WRAPPED : str
-        Represents a type of hay storage.
-    PROTECTED_TARPED : str
-        Represents a type of hay storage.
-    UNPROTECTED : str
-        Represents a type of hay storage.
-    BALEAGE : str
-        Represents a type of baleage storage.
-    DRY : str
-        Represents a type of grain storage.
-    HIGH_MOISTURE : str
-        Represents a type of grain storage.
-    BUNKER : str
-        Represents a type of silage storage.
-    PILE : str
-        Represents a type of silage storage.
-    BAG : str
-        Represents a type of silage storage.
-    """
-
-    PROTECTED_INDOORS = "Protected Indoors"
-    PROTECTED_WRAPPED = "Protected Wrapped"
-    PROTECTED_TARPED = "Protected Tarped"
-    UNPROTECTED = "Unprotected"
-    BALEAGE = "Baleage"
-    DRY = "Dry"
-    HIGH_MOISTURE = "High Moisture"
-    BUNKER = "Bunker"
-    PILE = "Pile"
-    BAG = "Bag"
